Update on-anger.xhtml
Added the missing end quote at the end of 3.15, I double checked this with the PDF but the quote was missing on WikiSource, have added it there too. WikiSource rendition of this piece has got quite a number of errors, I'll be proofreading it soon and will make correction there if I see any.